requirements.txt

# --- Web frameworks ---
flask==3.0.0
streamlit>=1.33.0

# --- Core dependencies ---
python-dotenv>=1.0.1
werkzeug==3.0.0
pandas>=2.2.2
numpy<2.0

# --- Audio analysis (optional) ---
# NOTE: Temporarily excluded from default install to avoid Python version conflicts
# on macOS arm64. Install separately from requirements-audio.txt when needed.
# openai>=1.3.0
# pyannote.audio>=3.1.1
# torch>=2.1.0
# torchaudio>=2.1.0
# soundfile>=0.12.1
# librosa

# --- Video/facial expression analysis ---
opencv-python>=4.9.0.80
altair>=5.3,<6
pillow>=10.4.0
# Py-Feat for AU/emotion via Detector
# Latest pip release is <=0.6.2 as of now; keep <0.7 for compatibility
py-feat>=0.6.0,<0.7

# --- ML/analytics ---
# Pin SciPy below 1.14 because nltools/py-feat transitively import scipy.integrate.simps
# which was removed in SciPy 1.14.
scipy>=1.11,<1.14
scikit-learn>=1.5.2
transformers>=4.43.3
google-generativeai>=0.7.2

# --- Real-time streaming ---
streamlit-webrtc>=0.50.0
webrtcvad>=2.0.10
av>=10.0.0
